requirements.txt
flask
openai-whisper==20231117
